+++ matrix-test.c
@@ -4,7 +4,7 @@
 
 int main() {
     matrix *m1 = random2((int[]){3, 3});
-    float nm2[3][3] = {
+    double nm2[3][3] = {
         {1, 2, 3},
         {3, 2, 1},
         {4, 5, 1}
