+++ matrix.h
@@ -2,7 +2,7 @@
 #define MATRIX_H
 
 typedef struct matrix {
-    float **mat;
+    double **mat;
     int dims[2];
 } matrix;
 
@@ -26,7 +26,7 @@ matrix *mul(int n, matrix *mat1);
 
 matrix *add(matrix *mat1, matrix *mat2);
 
-matrix *matrix2(int dims[2], float vals[][dims[1]]);
+matrix *matrix2(int dims[2], double vals[][dims[1]]);
 
 void unalloc(matrix *mat);
